+++ uptimerobot/api/monitor.go
@@ -17,6 +17,18 @@ var monitorType = map[string]int{
 }
 var MonitorType = mapKeys(monitorType)
 
+var monitorPostType = map[string]int{
+	"key-value":    1,
+	"raw data": 	2,
+}
+var MonitorPostType = mapKeys(monitorPostType)
+
+var monitorPostContentType = map[string]int{
+	"text/html":    		0,
+	"application/json": 	1,
+}
+var MonitorPostContentType = mapKeys(monitorPostContentType)
+
 var monitorSubType = map[string]int{
 	"http":   1,
 	"https":  2,
@@ -41,6 +53,17 @@ var monitorKeywordType = map[string]int{
 	"not exists": 2,
 }
 var MonitorKeywordType = mapKeys(monitorKeywordType)
+
+var monitorHTTPMethod = map[string]int{
+	"HEAD": 	1,
+	"GET": 		2,
+	"POST": 	3,
+	"PUT": 		4,
+	"PATCH": 	5,
+	"DELETE": 	6,
+	"OPTIONS": 	7,
+}
+var MonitorHTTPMethod = mapKeys(monitorHTTPMethod)
 
 var monitorHTTPAuthType = map[string]int{
 	"basic":  1,
@@ -74,9 +97,14 @@ type Monitor struct {
 
 	IgnoreSSLErrors bool `json:"ignore_ssl_errors"`
 
-	CustomHTTPHeaders map[string]string
+	CustomHTTPHeaders map[string]string `json:"custom_http_headers"`
 
-	AlertContacts []MonitorAlertContact
+	AlertContacts []MonitorAlertContact `json:"alert_contacts"`
+
+	HTTPMethod string `json:"http_method"`
+	PostType string `json:"post_type"`
+	PostContentType string `json:"post_content_type"`
+	PostValue map[string]string `json:"post_value"`
 }
 
 func (client UptimeRobotApiClient) GetMonitor(id int) (m Monitor, err error) {
@@ -85,6 +113,8 @@ func (client UptimeRobotApiClient) GetMonitor(id int) (m Monitor, err error) {
 	data.Add("ssl", fmt.Sprintf("%d", 1))
 	data.Add("custom_http_headers", fmt.Sprintf("%d", 1))
 	data.Add("alert_contacts", fmt.Sprintf("%d", 1))
+	data.Add("http_request_details", "true")
+	data.Add("auth_type", "true")
 
 	body, err := client.MakeCall(
 		"getMonitors",
@@ -115,6 +145,7 @@ func (client UptimeRobotApiClient) GetMonitor(id int) (m Monitor, err error) {
 	m.Type = intToString(monitorType, int(monitor["type"].(float64)))
 	m.Status = intToString(monitorStatus, int(monitor["status"].(float64)))
 	m.Interval = int(monitor["interval"].(float64))
+	m.HTTPMethod = intToString(monitorHTTPMethod, int(monitor["http_method"].(float64)))
 
 	switch m.Type {
 	case "port":
@@ -144,6 +175,19 @@ func (client UptimeRobotApiClient) GetMonitor(id int) (m Monitor, err error) {
 		m.HTTPUsername = monitor["http_username"].(string)
 		m.HTTPPassword = monitor["http_password"].(string)
 		break
+	}
+
+	switch m.HTTPMethod {
+	case "POST":
+		// not returned by the API
+		// m.PostType = intToString(monitorPostType, int(monitor["post_type"].(float64)))
+		// m.PostContentType = intToString(monitorPostContentType, int(monitor["post_content_type"].(float64)))
+		// post value
+		postValue := make(map[string]string)
+		for k, v := range monitor["post_value"].(map[string]interface{}) {
+			postValue[k] = v.(string)
+		}
+		m.PostValue = postValue
 	}
 
 	ignoreSSLErrors := int(monitor["ssl"].(map[string]interface{})["ignore_errors"].(float64))
@@ -203,6 +247,11 @@ type MonitorCreateRequest struct {
 	AlertContacts []MonitorRequestAlertContact
 
 	CustomHTTPHeaders map[string]string
+
+	HTTPMethod string
+	PostType string
+	PostContentType string
+	PostValue map[string]string
 }
 
 func (client UptimeRobotApiClient) CreateMonitor(req MonitorCreateRequest) (m Monitor, err error) {
@@ -211,6 +260,8 @@ func (client UptimeRobotApiClient) CreateMonitor(req MonitorCreateRequest) (m Mo
 	data.Add("url", req.URL)
 	data.Add("type", fmt.Sprintf("%d", monitorType[req.Type]))
 	data.Add("interval", fmt.Sprintf("%d", req.Interval))
+	data.Add("http_method", fmt.Sprintf("%d", monitorHTTPMethod[req.HTTPMethod]))
+
 	switch req.Type {
 	case "port":
 		data.Add("sub_type", fmt.Sprintf("%d", monitorSubType[req.SubType]))
@@ -228,6 +279,18 @@ func (client UptimeRobotApiClient) CreateMonitor(req MonitorCreateRequest) (m Mo
 		data.Add("http_auth_type", fmt.Sprintf("%d", monitorHTTPAuthType[req.HTTPAuthType]))
 		data.Add("http_username", req.HTTPUsername)
 		data.Add("http_password", req.HTTPPassword)
+		break
+	}
+
+	switch req.HTTPMethod {
+	case "POST":
+		data.Add("post_type", fmt.Sprintf("%d", monitorPostType[req.PostType]))
+		data.Add("post_content_type", fmt.Sprintf("%d", monitorPostContentType[req.PostContentType]))
+		// post value
+		jsonData, err := json.Marshal(req.PostValue)
+		if err == nil {
+			data.Add("post_value", string(jsonData))
+		}
 		break
 	}
 
@@ -287,6 +350,11 @@ type MonitorUpdateRequest struct {
 	AlertContacts []MonitorRequestAlertContact
 
 	CustomHTTPHeaders map[string]string
+
+	HTTPMethod string
+	PostType string
+	PostContentType string
+	PostValue map[string]string
 }
 
 func (client UptimeRobotApiClient) UpdateMonitor(req MonitorUpdateRequest) (m Monitor, err error) {
@@ -296,6 +364,8 @@ func (client UptimeRobotApiClient) UpdateMonitor(req MonitorUpdateRequest) (m Mo
 	data.Add("url", req.URL)
 	data.Add("type", fmt.Sprintf("%d", monitorType[req.Type]))
 	data.Add("interval", fmt.Sprintf("%d", req.Interval))
+	data.Add("http_method", fmt.Sprintf("%d", monitorHTTPMethod[req.HTTPMethod]))
+
 	switch req.Type {
 	case "port":
 		data.Add("sub_type", fmt.Sprintf("%d", monitorSubType[req.SubType]))
@@ -313,6 +383,18 @@ func (client UptimeRobotApiClient) UpdateMonitor(req MonitorUpdateRequest) (m Mo
 		data.Add("http_auth_type", fmt.Sprintf("%d", monitorHTTPAuthType[req.HTTPAuthType]))
 		data.Add("http_username", req.HTTPUsername)
 		data.Add("http_password", req.HTTPPassword)
+		break
+	}
+
+	switch req.HTTPMethod {
+	case "POST":
+		data.Add("post_type", fmt.Sprintf("%d", monitorPostType[req.PostType]))
+		data.Add("post_content_type", fmt.Sprintf("%d", monitorPostContentType[req.PostContentType]))
+		// post value
+		jsonData, err := json.Marshal(req.PostValue)
+		if err == nil {
+			data.Add("post_value", string(jsonData))
+		}
 		break
 	}
 
@@ -335,7 +417,7 @@ func (client UptimeRobotApiClient) UpdateMonitor(req MonitorUpdateRequest) (m Mo
 			data.Add("custom_http_headers", string(jsonData))
 		}
 	} else {
-		//delete custom http headers when it is empty
+		// delete custom http headers when it is empty
 		data.Add("custom_http_headers", "{}")
 	}